Brought table filter demo script closer in functionality to our imagined pipeline/dataset explorer. UI options not really fully considered at this point. Focussing on functionality.
Also re-fixed a bug in pipeline_add.js.

+++ shipyard/pipeline/static/pipeline/pipeline_add.js
@@ -154,13 +154,10 @@ $(document).ready(function(){ // wait for page to finish loading before executin
 
     $(document).on('keydown', function(e) {
         // backspace key also removes selected object
-        if (e.which === 8) {
-            if (!$(e.target).is("input, textarea")) {
-                // prevent backspace from triggering browser to navigate back one page
-                e.preventDefault();
-            } else {
-                canvasState.deleteObject();
-            }
+        if (e.which === 8 && !$(e.target).is("input, textarea")) {
+            // prevent backspace from triggering browser to navigate back one page
+            e.preventDefault();
+            canvasState.deleteObject();
         }
     })
